refactor(workbench): extract composer send path into useComposerSend
executeSend, submit, and post-submit draft clear live in the hook so send-path changes do not require opening AppWorkbench internals. Host still owns connect, live map, and settings.
