+++ src/app/AppWorkbench.tsx
@@ -93,10 +93,7 @@ import {
   applyTurnMarker,
   canSend,
   canType,
-  clearPriorTurnStreaming,
   isSessionLiveStreaming,
-  isSessionNotLiveError,
-  isTurnCancelledError,
   presentErrorBanner,
   type ErrorBannerView,
   weaveToolsIntoAssistantSegments,
@@ -111,7 +108,6 @@ import {
   type AskUserPayload,
   type ChatMessage,
   type PermissionPayload,
-  type SessionSnapshot,
 } from "@/lib/session";
 import {
   DEFAULT_SESSION_DATA_MODE,
@@ -204,8 +200,6 @@ import {
 } from "@/lib/escapeStop";
 import {
   isSameView,
-  isViewingSendTarget,
-  resolveComposerSendSessionId,
   shouldAdoptView,
 } from "@/lib/viewFocus";
 import {
@@ -417,7 +411,6 @@ import {
   composerSteerLive,
   resolveComposerSubmitAction,
 } from "@/lib/composerSendKey";
-import { countDraftChars } from "@/lib/draftStats";
 import {
   composerDraftStore,
   getDraft as getComposerDraft,
@@ -431,15 +424,9 @@ import {
   type ComposerProjectDraft,
 } from "@/lib/composerProjectDraft";
 import {
-  clearComposerSessionDraft,
   loadComposerSessionDraft,
   saveComposerSessionDraft,
 } from "@/lib/composerSessionDraft";
-import {
-  nextComposerSubmitSettlement,
-  shouldClearMatchingProjectDraft,
-  shouldClearProjectDraftAfterNewChatSend,
-} from "@/lib/composerSubmitClear";
 import {
   appendQuotesToContent,
   makeComposerQuoteId,
@@ -455,7 +442,6 @@ import {
   migrateDraftSendClaim,
   queueSessionKey,
   resolveSendQueueStripState,
-  shouldEnqueueSend,
   type QueuedSend,
 } from "@/lib/sendQueue";
 import {
@@ -474,7 +460,6 @@ import {
   type SlashItem,
 } from "@/lib/slashCatalog";
 import {
-  classifyWorkflowSlashLine,
   leftoverWorkflowArgs,
   resolveWorkflowSlashAction,
   stripWorkflowSlashFromDraft,
@@ -581,12 +566,9 @@ import { automationsBackgroundStatus } from "@/lib/automationsBackgroundStatus";
 import { recordAutomationRun } from "@/lib/automationRunHistory";
 import {
   extractAutomationPayload,
-  looksLikeScheduleIntent,
-  looksLikeScheduleUpdateIntent,
   recentUserPlainText,
   resolveAutomationUpsertTarget,
   shouldAutoApplyAutomationFence,
-  wrapAutomationSetupAgentText,
 } from "@/lib/automationSetup";
 
 import type { ComposerModelPick } from "@/lib/composerModelGroups";
@@ -735,6 +717,7 @@ import { useSearchPalette } from "@/hooks/useSearchPalette";
 import { useCompactDialog } from "@/hooks/useCompactDialog";
 import { useQueueEditDialog } from "@/hooks/useQueueEditDialog";
 import { useVoiceDictation } from "@/hooks/useVoiceDictation";
+import { useComposerSend } from "@/hooks/useComposerSend";
 import { useSessionCatalog } from "@/hooks/useSessionCatalog";
 import {
   createSessionNavHost,
@@ -1247,6 +1230,8 @@ export function AppWorkbench() {
    * `sessionSend` await cannot re-apply liveMap busy after UI already healed.
    */
   const sendEpochRef = useRef(0);
+  const sendQueueRef = useRef<ReturnType<typeof useSendQueue> | null>(null);
+  const showToastRef = useRef<(msg: string, ms?: number) => void>(() => {});
   const executeSendFromQueueRef = useRef<ExecuteSendFromQueue>(
     async () => false,
   );
@@ -6325,687 +6310,81 @@ export function AppWorkbench() {
     !editSubmitting &&
     !rewindBusy;
 
-  /**
-   * Dispatch one user turn (optimistic UI + connect + session_send).
-   * @param targetSessionId When set (queue flush), bind optimistic UI to this id.
-   * @param fromQueue Drop user+assistant on failure so requeue does not duplicate.
-   */
-  const executeSend = async (opts: {
-    storedDisplay: string;
-    att: Attachment[];
-    quotes?: ComposerQuote[];
-    goalMode: boolean;
-    fromQueue?: boolean;
-    targetSessionId?: string | null;
-    agentTextOverride?: string;
-  }): Promise<boolean> => {
-    // Session-keyed pool: secondary may send via shared Host (session-targeted).
-    if (!canLiveParticipate(isSecondaryWindowRef.current)) {
-      setLocalError(tr("session.secondaryLiveBanner"));
-      return false;
-    }
-    const { storedDisplay, att, goalMode: useGoal, fromQueue } = opts;
-    const quotesForSend = opts.quotes ?? [];
-    if (!fromQueue) await effortApplyRef.current;
-    const segments = parseStoredContent(storedDisplay);
-    if (isDraftEmpty(segments) && !att.length && !quotesForSend.length) {
-      sendInFlightRef.current = false;
-      return false;
-    }
-    const journalDisplay = appendQuotesToContent(storedDisplay, quotesForSend);
-    // Resolve the target before claiming so the lock is scoped to the chat
-    // that will actually receive the prompt (drafts use a stable sentinel).
-    const sendTargetId =
-      opts.targetSessionId !== undefined
-        ? opts.targetSessionId
-        : resolveComposerSendSessionId({
-            viewingSessionId: viewingSessionIdRef.current,
-            shellSessionId: session.sessionId,
-          });
-    const sendKey = queueSessionKey(sendTargetId);
-    if (!claimSendForSession(sendTargetId)) return false;
-    const heldSendKeys = new Set<string>([sendKey]);
-    const sendEpoch = (sendEpochBySessionRef.current.get(sendKey) ?? 0) + 1;
-    sendEpochBySessionRef.current.set(sendKey, sendEpoch);
-    sendEpochRef.current += 1;
-    const sendEpochCurrent = () =>
-      [...heldSendKeys].every(
-        (key) => sendEpochBySessionRef.current.get(key) === sendEpoch,
-      );
-    // Prefer viewing id over shell sessionId — openSession points viewing at
-    // the new chat before journal load finishes setSession; using only shell
-    // mis-routed sends into the previous (often stuck) chat.
-    const cacheKey = sendTargetId ?? "__draft__";
-    // Draft sends have no id to compare, so pin them to the view they came from:
-    // otherwise the optimistic bubbles / streaming state paint whatever *new*
-    // draft the user opened in the meantime.
-    const originView = currentViewFocus();
-    const viewingTarget = () =>
-      isViewingSendTarget(originView, currentViewFocus(), sendTargetId);
-
-    const agentBody = serializeQuotesForAgent(
-      quotesForSend,
-      serializeForAgent(segments, { goalMode: useGoal }),
-    );
-    let agentText = opts.agentTextOverride?.trim()
-      ? opts.agentTextOverride
-      : buildAgentPrompt(agentBody, att);
-    const schemaForSend = sessionJsonSchemaRef.current?.trim() || "";
-    if (schemaForSend && isActiveJsonSchema(schemaForSend)) {
-      agentText = wrapAgentTextWithJsonSchema(agentText, schemaForSend);
-    }
-    // Intent from user-visible body only (not /goal prefix or attachment paths).
-    const userIntentText = serializeForAgent(segments).trim();
-    const explicitAutomationSticky =
-      automationSetupDraftRef.current ||
-      (!!sendTargetId &&
-        automationSetupSessionsRef.current.has(sendTargetId));
-    // Goal mode is finite objective work — do not silently enter schedule setup
-    // unless the user opened “用 AI 创建” (sticky) for this session.
-    const scheduleIntent =
-      !useGoal &&
-      (looksLikeScheduleIntent(userIntentText) ||
-        looksLikeScheduleUpdateIntent(userIntentText));
-    const inAutomationSetup = explicitAutomationSticky || scheduleIntent;
-    if (inAutomationSetup) {
-      agentText = wrapAutomationSetupAgentText(agentText);
-    }
-    const titleSeed =
-      serializeForAgent(segments).replace(/\n/g, " ").trim() ||
-      quotesForSend[0]?.text.replace(/\n/g, " ").trim() ||
-      att.map((a) => a.name).join(", ");
-    const shouldAutoTitle =
-      isPlaceholderTitle(session.title) || !sendTargetId;
-    const ts = Date.now();
-    const userMessageId = `u-${ts}`;
-    const pendingAssistantId = `a-pending-${ts}`;
-    const dropIds = fromQueue
-      ? new Set([userMessageId, pendingAssistantId])
-      : new Set([pendingAssistantId]);
-    const stripOptimistic = (m: ChatMessage[]) =>
-      m.filter((x) => !dropIds.has(x.id));
-
-    if (editingUserMessageId) {
-      setEditingUserMessageId(null);
-      setEditAttachments([]);
-    }
-
-    if (viewingTarget()) setRetryStatus(null);
-    const nowIso = new Date().toISOString();
-    const appendOptimistic = (m: ChatMessage[]): ChatMessage[] => {
-      const cleaned = clearPriorTurnStreaming(m);
-      return [
-        ...cleaned,
-        {
-          id: userMessageId,
-          role: "user",
-          content: journalDisplay,
-          attachments: att.length ? att : undefined,
-          createdAt: nowIso,
-        },
-        {
-          id: pendingAssistantId,
-          role: "assistant",
-          content: "",
-          streaming: true,
-          createdAt: nowIso,
-        },
-      ];
-    };
-    if (sendTargetId) {
-      patchSessionMessages(sendTargetId, appendOptimistic);
-    } else if (viewingTarget()) {
-      setMessages((m) => {
-        const next = appendOptimistic(m);
-        messagesBySessionRef.current.set(cacheKey, next);
-        return next;
-      });
-    } else {
-      const prev = messagesBySessionRef.current.get(cacheKey) ?? [];
-      messagesBySessionRef.current.set(cacheKey, appendOptimistic(prev));
-    }
-    if (viewingTarget()) {
-      setSession((prev) =>
-        prev.state === "streaming" || prev.state === "awaiting_permission"
-          ? prev
-          : { ...prev, state: "streaming", lastError: null },
-      );
-      restartTurnClock(
-        resolveTurnClockKey(
-          sendTargetId ?? viewingSessionIdRef.current,
-        ),
-      );
-    }
-    // Optimistic liveHost only when we already own the live slot (or nothing is live).
-    // Never stamp streaming onto a foreign mid-turn — ensureConnected demotes first.
-    setLiveHost((prev) => {
-      if (prev.sessionId) {
-        if (sendTargetId && prev.sessionId !== sendTargetId) return prev;
-        // Draft / null target while another session is live → leave Host alone.
-        if (!sendTargetId && prev.sessionId) return prev;
-      }
-      const next = {
-        ...prev,
-        sessionId: sendTargetId ?? prev.sessionId,
-        state: "streaming" as const,
-        lastError: null,
-      };
-      liveHostRef.current = next;
-      return next;
-    });
-
-    const failStrip = () => {
-      if (sendTargetId) {
-        patchSessionMessages(sendTargetId, stripOptimistic);
-      } else {
-        const draftMsgs = messagesBySessionRef.current.get("__draft__");
-        if (draftMsgs) {
-          messagesBySessionRef.current.set(
-            "__draft__",
-            stripOptimistic(draftMsgs),
-          );
-        }
-        if (viewingTarget()) setMessages((m) => stripOptimistic(m));
-      }
-      if (viewingTarget()) {
-        setSession((prev) =>
-          prev.state === "streaming"
-            ? { ...prev, state: prev.sessionId ? "ready" : prev.state }
-            : prev,
-        );
-      }
-      // Symmetric rollback of optimistic liveHost streaming — otherwise
-      // useSendQueue.flush sees streaming forever and auto-flush starves.
-      // Mirror the optimistic guard: never rewind a foreign mid-turn we did not claim.
-      setLiveHost((prev) => {
-        if (prev.sessionId) {
-          if (sendTargetId && prev.sessionId !== sendTargetId) return prev;
-          if (!sendTargetId && prev.sessionId) return prev;
-        }
-        if (prev.state !== "streaming") return prev;
-        const next = {
-          ...prev,
-          state: (prev.sessionId ? "ready" : "idle") as SessionSnapshot["state"],
-        };
-        liveHostRef.current = next;
-        return next;
-      });
-    };
-
-    try {
-      let sessionId: string | null = null;
-      const live = liveHostRef.current;
-      if (
-        sendTargetId &&
-        live.sessionId === sendTargetId &&
-        live.state === "ready" &&
-        !live.lastError
-      ) {
-        sessionId = sendTargetId;
-      } else if (
-        fromQueue &&
-        sendTargetId &&
-        viewingSessionIdRef.current !== sendTargetId
-      ) {
-        failStrip();
-        return false;
-      } else {
-        sessionId = await ensureConnected({ sessionId: sendTargetId });
-      }
-      if (!sessionId) {
-        failStrip();
-        return false;
-      }
-      // Draft sends materialize a real id during ensureConnected. Atomically
-      // migrate the claim so a second call targeting the new id cannot slip
-      // through, while a newly opened draft remains independent.
-      const materializedSendKey = queueSessionKey(sessionId);
-      if (!heldSendKeys.has(materializedSendKey)) {
-        const claims = sendInFlightBySessionRef.current;
-        const alreadyOurs =
-          claims.has(materializedSendKey) &&
-          sendEpochBySessionRef.current.get(materializedSendKey) ===
-            sendEpoch;
-        if (claims.has(materializedSendKey) && !alreadyOurs) {
-          failStrip();
-          return false;
-        }
-        claims.delete(sendKey);
-        heldSendKeys.delete(sendKey);
-        if (sendEpochBySessionRef.current.get(sendKey) === sendEpoch) {
-          sendEpochBySessionRef.current.delete(sendKey);
-        }
-        if (!alreadyOurs) {
-          claims.add(materializedSendKey);
-          sendEpochBySessionRef.current.set(materializedSendKey, sendEpoch);
-        }
-        heldSendKeys.add(materializedSendKey);
-      }
-      if (fromQueue && sendTargetId && sessionId !== sendTargetId) {
-        failStrip();
-        return false;
-      }
-      // Bind draft message cache to the real id early (Host already materialized).
-      // Queue migrate waits until sessionSend succeeds so a failed flush can
-      // requeue under the original claim key (`__draft__`) without splitting.
-      if (!sendTargetId) {
-        const draftMsgs = messagesBySessionRef.current.get("__draft__");
-        if (draftMsgs?.length) {
-          messagesBySessionRef.current.set(sessionId, draftMsgs);
-          messagesBySessionRef.current.delete("__draft__");
-        }
-        if (migrateDraftTurnClock(turnStartedAtBySessionRef.current, sessionId)) {
-          syncViewedTurnClock(sessionId);
-        }
-      }
-      // Sticky setup only for explicit “用 AI 创建”, so a one-off “每天…” line
-      // does not force every later message in the chat into automation mode.
-      if (automationSetupDraftRef.current) {
-        automationSetupSessionsRef.current.add(sessionId);
-        automationSetupDraftRef.current = false;
-      }
-      if (
-        fromQueue &&
-        sendTargetId &&
-        liveHostRef.current.sessionId &&
-        liveHostRef.current.sessionId !== sendTargetId
-      ) {
-        failStrip();
-        return false;
-      }
-      // Bind the turn to `sessionId`, never to "whatever is live". Host
-      // re-focuses that chat (background/parked → live) before prompting, so a
-      // warm connect racing this send cannot deliver it to another chat — and
-      // a mid-send "new chat" still lets this turn complete.
-      try {
-        await api.sessionSend(agentText, journalDisplay, sessionId, att);
-      } catch (sendErr) {
-        // Stop / stall during Host vision/prepare: prompt never left.
-        // Do not retry and do not treat as CONNECT_FAILED (P0-2).
-        if (isTurnCancelledError(sendErr)) {
-          // Journal already has the user row; drop only the empty assistant shell.
-          if (sendTargetId) {
-            patchSessionMessages(sendTargetId, (m) =>
-              m.filter((x) => x.id !== pendingAssistantId),
-            );
-          } else if (viewingTarget()) {
-            setMessages((m) => m.filter((x) => x.id !== pendingAssistantId));
-          }
-          if (viewingTarget()) {
-            setSession((prev) =>
-              prev.state === "streaming"
-                ? { ...prev, state: prev.sessionId ? "ready" : prev.state }
-                : prev,
-            );
-          }
-          setLiveMap((prev) =>
-            projectHostIntoLiveMap(prev, {
-              sessionId,
-              state: "ready",
-              streamingMessageId: null,
-            }),
-          );
-          return true;
-        }
-        // Host refuses rather than misroute when the chat lost its process
-        // (idle recycle / crash while `liveHost` still looked ready).
-        // Cold-connect that chat once, then retry the same turn.
-        if (!isSessionNotLiveError(sendErr)) throw sendErr;
-        if (!sendEpochCurrent()) return false;
-        const reconnected = await ensureConnected({
-          sessionId,
-          force: true,
-        });
-        if (reconnected !== sessionId) throw sendErr;
-        if (!sendEpochCurrent()) return false;
-        await api.sessionSend(agentText, journalDisplay, sessionId, att);
-      }
-      // Ghost heal / newer send superseded this await — do not re-dirty UI.
-      if (!sendEpochCurrent()) return false;
-      // Keep liveMap busy for this session if the user already left the thread.
-      setLiveMap((prev) =>
-        projectHostIntoLiveMap(prev, {
-          sessionId,
-          state: "streaming",
-          streamingMessageId: null,
-        }),
-      );
-      // Only after a successful send: move remaining draft follow-ups onto the
-      // real session. If this threw, claim requeues under `__draft__` intact.
-      if (!sendTargetId) {
-        sendQueue.migrateDraft(sessionId);
-      }
-      // Cross-session recent prompts (localStorage ring, max 50).
-      // Store display form so chips/skill tokens rehydrate in the composer.
-      if (storedDisplay.trim()) {
-        setRecentPromptHistory(
-          recordRecentPrompt({
-            text: storedDisplay,
-            sessionId,
-            at: nowIso,
-          }),
-        );
-      }
-      // `session.autoTitle` is on the mirror allowlist, so phone chats get a
-      // real title instead of staying on the "new chat" placeholder forever.
-      if (shouldAutoTitle && api.hasHost()) {
-        void api
-          .sessionAutoTitle(sessionId, titleSeed)
-          .then((meta) => {
-            if (meta?.title) applySessionTitle(sessionId, meta.title);
-          })
-          .catch(() => {
-            /* ignore */
-          });
-      }
-      return true;
-    } catch (e) {
-      if (!sendEpochCurrent()) return false;
-      failStrip();
-      if (viewingTarget()) setLocalError(String(e));
-      return false;
-    } finally {
-      for (const key of heldSendKeys) {
-        // A ghost-heal or newer send may already own this session key. An old
-        // await must not release that newer claim when it finally settles.
-        if (sendEpochBySessionRef.current.get(key) !== sendEpoch) continue;
-        sendInFlightBySessionRef.current.delete(key);
-        sendEpochBySessionRef.current.delete(key);
-      }
-      sendInFlightRef.current = sendInFlightBySessionRef.current.size > 0;
-    }
-  };
-
-  const persistComposerSubmitClear = (opts?: {
-    clearProjectDraft?: boolean;
-    clearSessionDraft?: boolean;
-    sessionDraftId?: string | null;
-    sentText?: string;
-    sentAttachments?: Attachment[];
-    sentQuotes?: ComposerQuote[];
-  }) => {
-    const projectKey = projectDraftKey(activeProject?.id ?? null);
-    const savedProjectDraft = loadComposerProjectDraft(projectKey);
-    if (opts?.clearProjectDraft) {
-      clearComposerProjectDraft(projectKey);
-    } else if (
-      shouldClearMatchingProjectDraft({
-        projectDraftText: savedProjectDraft?.text,
-        projectDraftAttachments: savedProjectDraft?.attachments,
-        projectDraftQuotes: savedProjectDraft?.quotes,
-        sentText: opts?.sentText ?? "",
-        sentAttachments: opts?.sentAttachments,
-        sentQuotes: opts?.sentQuotes,
-      })
-    ) {
-      clearComposerProjectDraft(projectKey);
-    }
-    if (opts?.clearSessionDraft) {
-      const sid =
-        opts.sessionDraftId ??
-        viewingSessionIdRef.current ??
-        session.sessionId ??
-        null;
-      if (sid) clearComposerSessionDraft(sid);
-    }
-  };
-
-  /** Wipe the visible composer now. Persist is a separate call after send settles. */
-  const resetComposerUiAfterSubmit = () => {
-    setDraft("");
-    setQuotes([]);
-    promptHistoryIndexRef.current = null;
-    setPromptHistoryIndex(null);
-    setPromptHistoryOpen(false);
-    setPromptHistoryFilter("");
-    setPromptHistoryActive(0);
-    setPromptHistoryFocusFilter(false);
-    setPromptHistoryScope("session");
-    setSlashQuery(null);
-    setAttachments([]);
-    setChatAttachments([]);
-    setAttachChatOpen(false);
-    requestAnimationFrame(() => {
-      const el = composerInputRef.current;
-      if (el) el.style.height = "auto";
-    });
-  };
-
-  const clearComposerAfterSubmit = (opts?: {
-    /** Drop the per-project new-chat buffer (only when leaving a draft send). */
-    clearProjectDraft?: boolean;
-    /** Drop the per-session follow-up buffer (send / clear on a real thread). */
-    clearSessionDraft?: boolean;
-    sessionDraftId?: string | null;
-  }) => {
-    resetComposerUiAfterSubmit();
-    persistComposerSubmitClear(opts);
-  };
-
-  /**
-   * Wipe the main composer (text + attachments). Also leaves inline edit mode
-   * and drops the per-project new-chat buffer when on a draft page, or the
-   * per-session follow-up buffer when on a real thread.
-   */
-  const applyClearComposerDraft = useCallback(() => {
-    const onDraftPage =
-      session.sessionId == null && viewingSessionIdRef.current == null;
-    setQuotes([]);
-    clearComposerAfterSubmit({
-      clearProjectDraft: onDraftPage,
-      clearSessionDraft: !onDraftPage,
-    });
-    if (!editSubmitting) {
-      setEditingUserMessageId(null);
-      setEditAttachments([]);
-    }
-    requestComposerFocus();
-  }, [editSubmitting, requestComposerFocus, session.sessionId]);
-
-  /** Clear immediately, or confirm first when the draft is long (>200 chars). */
-  const requestClearComposerDraft = useCallback(() => {
-    const draft = getDraft();
-    const hasBody =
-      !isDraftEmpty(parseStoredContent(draft)) ||
-      attachments.length > 0 ||
-      chatAttachments.length > 0 ||
-      quotes.length > 0;
-    if (!hasBody) return;
-    if (countDraftChars(draft) > 200) {
-      setAppDialog({
-        kind: "confirm",
-        title: tr("composer.clearDraftConfirmTitle"),
-        message: tr("composer.clearDraftConfirmMessage"),
-        confirmLabel: tr("composer.clearDraftConfirm"),
-        danger: true,
-        onConfirm: () => applyClearComposerDraft(),
-      });
-      return;
-    }
-    applyClearComposerDraft();
-  }, [applyClearComposerDraft, attachments.length, chatAttachments.length, quotes.length, getDraft, tr]);
-
-  /** Enqueue when agent is busy; otherwise send immediately. */
-  const send = async () => {
-    if (!canLiveParticipate(isSecondaryWindowRef.current)) {
-      showToast(tr("session.secondaryLiveBanner"), 4000);
-      return;
-    }
-    // Flush live contenteditable → draft before send so the last Enter / blank
-    // lines (and any debounced newline commit) are not dropped from the bubble.
-    const editorEl = composerInputRef.current;
-    let draft = getDraft();
-    if (editorEl) {
-      try {
-        const live = serializeDom(editorEl);
-        if (live !== draft) {
-          setDraft(live);
-          draft = live;
-        }
-      } catch {
-        /* keep getDraft() */
-      }
-    }
-    const fromDraft = parseChatTokens(draft, (id) =>
-      lookupChatTitle(id, sessions, ""),
-    );
-    let refs = chatAttachments;
-    for (const extra of fromDraft) {
-      refs = addChatRef(refs, extra, { currentId: session.sessionId }).refs;
-    }
-    const storedDisplay = prependChatTokens(stripChatTokens(draft), refs);
-    const segments = parseStoredContent(storedDisplay);
-    const att = attachments;
-    const sendQuotes = quotesRef.current;
-    if (isDraftEmpty(segments) && !att.length && !sendQuotes.length) return;
-    // Lone /workflow(s) — App has no TUI dashboard. Bare command opens Settings.
-    // `/workflow <args>` falls through as a normal session turn.
-    if (!att.length && !segments.some((s) => s.type === "skill")) {
-      const plain = segments
-        .map((s) => (s.type === "text" ? s.text : ""))
-        .join("");
-      const wf = classifyWorkflowSlashLine(plain);
-      if (wf?.kind === "dashboard") {
-        clearComposerAfterSubmit({
-          clearProjectDraft: session.sessionId == null,
-          clearSessionDraft: session.sessionId != null,
-          sessionDraftId: viewingSessionIdRef.current ?? session.sessionId,
-        });
-        openWorkflowsSettings();
-        return;
-      }
-    }
-    if (session.state === "awaiting_permission") {
-      showToast(tr("composer.queueBlockedPermission"), 2800);
-      return;
-    }
-    // Unassigned chats use workspaces/general as cwd (no sidebar project).
-    sendQueue.releaseFlushHold();
-
-    // New-chat page → after send, forget the project buffer so restore is empty.
-    // Existing-session follow-ups clear the per-session buffer only (not project).
-    const fromNewChatPage = session.sessionId == null;
-    const clearDraftOpts = {
-      clearProjectDraft: fromNewChatPage,
-      clearSessionDraft: !fromNewChatPage,
-      sessionDraftId: viewingSessionIdRef.current ?? session.sessionId,
-      sentText: storedDisplay,
-      sentAttachments: att,
-      sentQuotes: sendQuotes,
-    };
-
-    // Enqueue only when *this viewed chat* FSM is busy (streaming/connecting).
-    // Host mid-turn on another session → executeSend demotes + spawns concurrent
-    // work. Never park a new-chat / other-session send into a fake local queue
-    // (that showed “本会话队列” on empty welcome while the real turn ran elsewhere).
-    // Also ignore the process-global `connecting` flag — foreign ensureConnected
-    // must not make SuperGrok welcome enqueue (see shouldEnqueueSend).
-    if (shouldEnqueueSend(session.state, connecting)) {
-      sendQueue.enqueue({
-        storedDisplay,
-        attachments: att,
-        quotes: sendQuotes,
-        goalMode,
-      });
-      clearComposerAfterSubmit(clearDraftOpts);
-      return;
-    }
-
-    // Clear the box with the optimistic user bubble — do not wait for
-    // ensureConnected / sessionSend (that left the prompt sitting in the
-    // composer for seconds). Persist + fail-restore settle after executeSend.
-    const originView = currentViewFocus();
-    resetComposerUiAfterSubmit();
-
-    const sent = await executeSend({
-      storedDisplay,
-      att,
-      quotes: sendQuotes,
-      goalMode,
-    });
-    if (sent && refs.length) {
-      showToast(
-        tr("attachChat.sentWith", {
-          titles: refs
-            .map((r) => r.title.trim() || r.sessionId.slice(0, 8))
-            .join(" · "),
-        }),
-        2600,
-      );
-    }
-    const action = nextComposerSubmitSettlement({
-      sendSucceeded: sent,
-      sentText: storedDisplay,
-      sentAttachments: att,
-      sentQuotes: sendQuotes,
-      currentText: getDraft(),
-      currentAttachments: attachmentsRef.current,
-      currentQuotes: quotesRef.current,
-    });
-    const sendTargetId = resolveComposerSendSessionId({
-      viewingSessionId: originView.sessionId,
-      shellSessionId: session.sessionId,
-    });
-    const stillHere = isViewingSendTarget(
-      originView,
-      currentViewFocus(),
-      sendTargetId,
-    );
-    if (stillHere) {
-      if (action === "persist-clear") persistComposerSubmitClear(clearDraftOpts);
-      else if (action === "restore") {
-        setDraft(storedDisplay);
-        setAttachments(att);
-        setChatAttachments(refs);
-        setQuotes(sendQuotes);
-      }
-      return;
-    }
-    // Navigated away: leaving already persisted the composer-at-leave
-    // (follow-up text stays). Only refill an empty origin buffer on fail.
-    // New-chat send adopts the materialized session, so stillHere is false —
-    // still wipe the per-project new-session buffer or the next "New session"
-    // restores the just-sent prompt (#620).
-    if (sent) {
-      persistComposerSubmitClear({
-        clearProjectDraft: shouldClearProjectDraftAfterNewChatSend({
-          fromNewChatPage: clearDraftOpts.clearProjectDraft,
-          sendSucceeded: true,
-        }),
-        sentText: storedDisplay,
-        sentAttachments: att,
-        sentQuotes: sendQuotes,
-      });
-      return;
-    }
-    if (clearDraftOpts.clearSessionDraft && clearDraftOpts.sessionDraftId) {
-      if (!loadComposerSessionDraft(clearDraftOpts.sessionDraftId)) {
-        saveComposerSessionDraft(clearDraftOpts.sessionDraftId, {
-          text: storedDisplay,
-          attachments: att,
-          chatAttachments: refs,
-          quotes: sendQuotes,
-          goalMode,
-        });
-      }
-    } else if (
-      clearDraftOpts.clearProjectDraft &&
-      !loadComposerProjectDraft(projectDraftKey(activeProject?.id ?? null))
-    ) {
-      saveComposerProjectDraft(projectDraftKey(activeProject?.id ?? null), {
-        text: storedDisplay,
-        attachments: att,
-        chatAttachments: refs,
-        quotes: sendQuotes,
-        goalMode,
-      });
-    }
-  };
-  sendRef.current = send;
+  const {
+    executeSend,
+    send,
+    clearComposerAfterSubmit,
+    requestClearComposerDraft,
+  } = useComposerSend({
+    tr,
+    session,
+    sessions,
+    activeProject,
+    connecting,
+    goalMode,
+    attachments,
+    chatAttachments,
+    quotes,
+    editSubmitting,
+    editingUserMessageId,
+    isPlaceholderTitle,
+    isSecondaryWindowRef,
+    viewingSessionIdRef,
+    composerInputRef,
+    liveHostRef,
+    messagesBySessionRef,
+    sessionJsonSchemaRef,
+    automationSetupDraftRef,
+    automationSetupSessionsRef,
+    sendInFlightRef,
+    sendInFlightBySessionRef,
+    sendEpochRef,
+    sendEpochBySessionRef,
+    turnStartedAtBySessionRef,
+    effortApplyRef,
+    promptHistoryIndexRef,
+    quotesRef,
+    attachmentsRef,
+    sendQueueRef,
+    showToastRef,
+    sendRef,
+    executeSendFromQueueRef,
+    executeSendLatestRef,
+    claimSendForSession,
+    currentViewFocus,
+    patchSessionMessages,
+    ensureConnected,
+    getDraft,
+    requestComposerFocus,
+    openWorkflowsSettings,
+    applySessionTitle,
+    restartTurnClock,
+    syncViewedTurnClock,
+    setLocalError,
+    setSession,
+    setMessages,
+    setLiveHost,
+    setLiveMap,
+    setDraft,
+    setQuotes,
+    setAttachments,
+    setChatAttachments,
+    setAttachChatOpen,
+    setSlashQuery,
+    setPromptHistoryIndex,
+    setPromptHistoryOpen,
+    setPromptHistoryFilter,
+    setPromptHistoryActive,
+    setPromptHistoryFocusFilter,
+    setPromptHistoryScope,
+    setEditingUserMessageId,
+    setEditAttachments,
+    setRetryStatus,
+    setRecentPromptHistory,
+    setAppDialog,
+  });
   voiceDictationAutoSendRef.current = voiceDictationAutoSend;
 
-  executeSendFromQueueRef.current = (opts) => executeSend(opts);
-  executeSendLatestRef.current = executeSend;
 
   const queuePreviewLabels = useMemo(
     () => ({
@@ -8251,6 +7630,7 @@ export function AppWorkbench() {
   }, []);
 
   voiceNotifyRef.current = showToast;
+  showToastRef.current = showToast;
 
   const promptCreateSpace = useCallback(
     (afterCreate?: (id: string) => void) => {
@@ -8889,6 +8269,8 @@ export function AppWorkbench() {
     acceptExternal: !isSecondaryWindow,
     labels: sendQueueLabels,
   });
+
+  sendQueueRef.current = sendQueue;
 
   const queueEdit = useQueueEditDialog({
     tr,
